pyproject.toml
[project]
name = "abstract-block-dumper"
requires-python = ">=3.11"
keywords = []
license = {text = "MIT"}
readme = "README.md"
authors = [
    {name = "Reef Technologies", email = "opensource@reef.pl"},
]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = [
    "version",
]
dependencies = [
    "bittensor>=9.10.1",
    "celery>=5.5.3",
    "django>=3.2,<6.0",
    "structlog>=25.4.0",
]

[project.urls]
"Source" = "https://github.com/bactensor/abstract-block-dumper"
"Issue Tracker" = "https://github.com/bactensor/abstract-block-dumper/issues"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"


[tool.hatch.build.hooks.vcs]
version-file = "src/abstract_block_dumper/_version.py"

[dependency-groups]
test = [
    "factory-boy",
    "freezegun",
    "pytest",
    "pytest-asyncio",
    "pytest-django",
    "pytest-xdist",
]
lint = [
    "codespell[toml]",
    "django-stubs>=1.12.0",
    "mypy>=1.8",
    "ruff",
    "types-freezegun",
    "types-python-dateutil",
    "types-requests",
]
release = [
    "towncrier",
]
dev = [
    "pre-commit",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]  # allow for `import tests` in test files
target_package_name = "abstract_block_dumper"  # required by pytest-apiver
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "D",
    "E", "F", "I", "UP",
    "TC005",
]
ignore = [
    "D100", "D105", "D106", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102", "D103", "D104", # TODO remove once we have docstring for all public methods
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test/**" = ["D", "F403", "F405"]

[tool.codespell]
skip = "*.min.js,uv.lock"
ignore-words-list = "datas"


[tool.towncrier]
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
title_format = "## [{version}](https://github.com/bactensor/abstract-block-dumper/releases/tag/v{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/bactensor/abstract-block-dumper/issues/{issue})"

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Doc"
showcontent = true

[[tool.towncrier.type]]
directory = "infrastructure"
name = "Infrastructure"
showcontent = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
strict_optional = true
exclude = [
    "tests/",
]

[tool.django-stubs]
django_settings_module = "tests.settings"

[[tool.mypy.overrides]]
module = [
    "bittensor",
    "bittensor.*",
    "celery",
    "celery.*",
    "cloudpickle",
    "cloudpickle.*",
    "nox",
    "pytest",
    "tests.*",
]
ignore_missing_imports = true
